fix: handle destructuring with mixed declarations and assignments
Fixes https://github.com/decaffeinate/decaffeinate/issues/472

In a case like `({a, b: c.d} = e);`, a `var` was being added to the front, but
that's not valid when some of the destructured values are property assignments.
To fix, I made it so we walk the LHS to see if there are any non-identifier
assignments, and fall back to the normal variable declaration case if so.