fix: always use magic-string's insertLeft to insert code
Fixes https://github.com/decaffeinate/decaffeinate/issues/521

The `insertRight` function in magic-string has the weird property that repeated
inserts at the same index show up in reverse order in the resulting string,
which caused a bug where a `var` was inserted on the wrong line. While there may
be a way to make `insertRight` work, the simpler approach seems to be to do the
same thing as decaffeinate: always use `insertLeft`, and try to do insert
operations in order.

In testing this, I also made sure that all decaffeinate tests pass with this
change applied.