fix: rework state management to better track declaration scopes (#42)
Fixes https://github.com/decaffeinate/decaffeinate/issues/605
Fixes https://github.com/decaffeinate/decaffeinate/issues/316
Fixes #8

Previously, the strategy was to add variable declarations as they're seen using
magic-string inserts. This approach had a few fundamental problems:
* We want to declare variables only once and in the most specific scope
  possible. That means that if we see `x = 5`, we don't know yet if it's
  used/assigned in a sibling or less specific scope, so we need to wait until
  the end of the function before we know for sure where the `x` declaration
  should go.
* Ideally if we see multiple declarations that need to be moved to the top of
  the block (e.g. `[a, b.c] = d` needs to declare `a` externally), we would
  join them all together into the same declaration, so at most one added per
  block.
* The TraverseState code never implemented function scopes, so in some cases,
  the "is this variable already declared?" logic would miss a variable. I added
  a test for this as assignment-to-declaration-nested-in-parent-scope.
* The code only observed assignments, not usages, which isn't enough information
  to declare variables properly when they are assigned in a more specific scope
  than they are used.

To implement all of this, I reworked the state to be smarter and changed the
general patching strategy so that we only actually patch in `var`s when we exit
the relevant function scope. For each variable, we keep track of a
`BindingState` value that tracks the most specific scope of all usages so far.
For any inline binding (e.g. `[a, b] = c`, where we could potentially prepend
with `var`), we keep track of it, and only put the `var` inline if none of the
variables ended up moving to a higher scope.

In addition to the main scope-tracking change, I also made all cases properly
use the `lhsHasNonIdentifierAssignment` function to correctly determine if `var`
can be safely inserted inline.

Another approach I considered was to do a two (or more) pass approach where we
precompute usages of different variables, but doing it in one pass ended up
being reasonable and better fits some semantics, like the case described in the
assignment-in-parent-after-nested-scope test.

This change breaks the decaffeinate tests by outputting better variable
declarations, so I'm marking it as a breaking change.

BREAKING CHANGE: The resulting source code writes declarations in different
places in some cases, generally in an improved way.