fix: use the proper scope for assignments within method keys
Progress toward https://github.com/decaffeinate/decaffeinate/issues/858

The babylon AST has ObjectMethod and ClassMethod keys that are within the method
node, which means that we were viewing any assignment within that key as being
in the method block, so the variable would get declared in the wrong place. To
work around this, we can temporarily set the scope as one higher while we
traverse the key part of the method.