feat: allow optional chaining, upgrade babel dependencies
Progress toward https://github.com/decaffeinate/decaffeinate/issues/336

Babel 7 now has initial support for optional chaining, and we need to enable it
in the parser for all decaffeinate dependencies in order for it to work at all.
There's a little risk in using a prerelease version of Babel (which also doesn't
have updated types yet), but it seems probably fine.

Also throw in a magic-string upgrade.