fix: more precisely detect functions in canOwnBindings (#79)
Fixes https://github.com/decaffeinate/decaffeinate/issues/1201

ArrowFunctionExpression was an example of a function node type that wasn't being
caught here, and it looks like there are also newer node types for private
functions and functions used in exports. The `isFunction` function in
babel-types is the more robust way to do this check, so I just switched to that.